chore(error): move variable cheks into action-function for all settings, and move CountiesSelect out of MemberAdministration
