+++ tavla/app/(admin)/organizations/[id]/page.tsx
@@ -3,7 +3,7 @@ import { Heading1, LeadParagraph } from '@entur/typography'
 import { redirect } from 'next/navigation'
 import { UploadLogo } from '../components/UploadLogo'
 import { MemberAdministration } from '../components/MemberAdministration'
-import { CountiesSelect } from '../components/MemberAdministration/CountiesSelect'
+import { CountiesSelect } from '../components/CountiesSelect'
 import { FontSelect } from '../components/FontSelect'
 import { DefaultColumns } from '../components/DefaultColumns'
 import { getOrganizationIfUserHasAccess } from 'app/(admin)/actions'
