+++ tavla/app/(admin)/organizations/components/CountiesSelect/actions.ts
@@ -1,4 +1,5 @@
 'use server'
+import { getFormFeedbackForError } from 'app/(admin)/utils'
 import {
     initializeAdminApp,
     userCanEditOrganization,
@@ -7,20 +8,22 @@ import { handleError } from 'app/(admin)/utils/handleError'
 import { firestore } from 'firebase-admin'
 import { revalidatePath } from 'next/cache'
 import { redirect } from 'next/navigation'
-import { TCountyID, TOrganizationID } from 'types/settings'
+import { TOrganizationID } from 'types/settings'
 
 initializeAdminApp()
 
 export async function setCounties(
-    oid: TOrganizationID,
-    countiesList: TCountyID[],
+    oid: TOrganizationID | undefined,
+    data: FormData,
 ) {
+    if (!oid) return getFormFeedbackForError()
     const access = userCanEditOrganization(oid)
     if (!access) return redirect('/')
+    const counties = data.getAll('county') as string[]
 
     try {
         await firestore().collection('organizations').doc(oid).update({
-            'defaults.counties': countiesList,
+            'defaults.counties': counties,
         })
         revalidatePath(`/organizations/${oid}`)
     } catch (e) {
